comment: add compressor comments
