+++ byteps/common/compressor/impl/dithering.cc
