+++ byteps/common/compressor/impl/onebit.cc
