+++ byteps/common/compressor/impl/vanilla_error_feedback.h
@@ -13,8 +13,8 @@
 // limitations under the License.
 // =============================================================================
 
-#ifndef BYTEPS_COMPRESSOR_STRATEGY_VANILLA_ERROR_FEEDBACK_H
-#define BYTEPS_COMPRESSOR_STRATEGY_VANILLA_ERROR_FEEDBACK_H
+#ifndef BYTEPS_COMPRESSOR_IMPL_VANILLA_ERROR_FEEDBACK_H
+#define BYTEPS_COMPRESSOR_IMPL_VANILLA_ERROR_FEEDBACK_H
 
 #include "../error_feedback.h"
 
@@ -24,18 +24,43 @@ namespace compressor {
 
 /*!
  * \brief VanillaErrorFeedbackCompressor
+ *
+ * paper: Communication-efficient distributed blockwise momentum sgd with
+ * error-feedback
+ * https://arxiv.org/pdf/1905.10936.pdf
+ * 
+ * each worker i:
+ *    p_{t,i} <- g_{t,i} + \frac{\eta_{t-1}}{\eta_t} e_{t,i}
+ *    c_{t,i} <- Compress(p_{t,i})
+ *    e_{t,i} <- p_{t,i} - c_{t,i}
+ * 
+ * server:
+ *    \tilde{p}_{t} <- \frac{1}{M} \sum_{i=1}^{M} c_{t,i} +\frac{\eta_{t-1}}{\eta_{t}} \tilde{e_t}
+ *    \tilde{e}_{t+1} <- \tilde{p}_{t}-\tilde{c_t}
+ * 
+ * Error-correction: error needs to be scaled with \frac{\eta_{t-1}}{\eta_t}.
  */
 class VanillaErrorFeedbackCompressor : public ErrorFeedback {
  public:
-  VanillaErrorFeedbackCompressor(size_t size,
-      std::unique_ptr<Compressor> cptr);
+  VanillaErrorFeedbackCompressor(size_t size, std::unique_ptr<Compressor> cptr);
   virtual ~VanillaErrorFeedbackCompressor();
 
  protected:
+  /*!
+   * \brief 
+   * 
+   * \param grad 
+   */
   void UpdateGradient(tensor_t grad) override;
-  
+
  private:
+  /*!
+   * \brief learning rate
+   * 
+   * read from file each step
+   */
   double _pre_lr, _cur_lr;
+
   int _fd;
   void* _mm;
 };
@@ -43,4 +68,4 @@ class VanillaErrorFeedbackCompressor : public ErrorFeedback {
 }  // namespace common
 }  // namespace byteps
 
-#endif  // BYTEPS_COMPRESSOR_STRATEGY_VANILLA_ERROR_FEEDBACK_H
+#endif  // BYTEPS_COMPRESSOR_IMPL_VANILLA_ERROR_FEEDBACK_H
